part-1/requirements.txt
langchain-openai
langchain-community
langchain-core
langchain
langgraph
typing-extensions
streamlit
requests
chromadb
ipykernel
python-dotenv